### 2024-08-22 #7128 @editorialbot
[PRE REVIEW]: dfitspy: a dfits and fitsort implementation in Python
**Submitting author:** @astrom-tom (<a href="http://orcid.org/0000-0001-8385-3276">Romain Thomas</a>)
**Repository:** https://github.com/Romain-Thomas-Shef/dfitspy_RSECon24
**Branch with paper.md** (empty if default branch):
**Version:** v20.4.1
