### 2024-08-22 #7129 @EdwinB12
I found a couple of issues, here they are:

https://github.com/Romain-Thomas-Shef/dfitspy_RSECon24/issues/2
