### 2024-08-22 #7129 @arfon
@EdwinB12 - Are you happy with the changes made by the author? If so, can you confirm you recommend this paper be published in JOSS?
