### 2024-08-22 #7129 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
